+++ private/buf/cmd/buf/command/registry/plugin/pluginupdate/pluginupdate.go
@@ -12,17 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package moduleupdate
+package pluginupdate
 
 import (
 	"context"
 	"fmt"
 
-	modulev1 "buf.build/gen/go/bufbuild/registry/protocolbuffers/go/buf/registry/module/v1"
+	pluginv1beta1 "buf.build/gen/go/bufbuild/registry/protocolbuffers/go/buf/registry/plugin/v1beta1"
 	"connectrpc.com/connect"
 	"github.com/bufbuild/buf/private/buf/bufcli"
 	"github.com/bufbuild/buf/private/bufpkg/bufparse"
-	"github.com/bufbuild/buf/private/bufpkg/bufregistryapi/bufregistryapimodule"
+	"github.com/bufbuild/buf/private/bufpkg/bufregistryapi/bufregistryapiplugin"
 	"github.com/bufbuild/buf/private/pkg/app/appcmd"
 	"github.com/bufbuild/buf/private/pkg/app/appext"
 	"github.com/bufbuild/buf/private/pkg/syserror"
@@ -30,18 +30,17 @@ import (
 )
 
 const (
-	visibilityFlagName   = "visibility"
-	descriptionFlagName  = "description"
-	urlFlagName          = "url"
-	defaultLabelFlagName = "default-label-name"
+	visibilityFlagName  = "visibility"
+	descriptionFlagName = "description"
+	urlFlagName         = "url"
 )
 
-// NewCommand returns a new Command
+// NewCommand returns a new Command.
 func NewCommand(name string, builder appext.SubCommandBuilder) *appcmd.Command {
 	flags := newFlags()
 	return &appcmd.Command{
-		Use:   name + " <remote/owner/module>",
-		Short: "Update BSR module settings",
+		Use:   name + " <remote/owner/plugin>",
+		Short: "Update BSR plugin settings",
 		Args:  appcmd.ExactArgs(1),
 		Run: builder.NewRunFunc(
 			func(ctx context.Context, container appext.Container) error {
@@ -69,19 +68,13 @@ func (f *flags) Bind(flagSet *pflag.FlagSet) {
 		flagSet,
 		descriptionFlagName,
 		&f.Description,
-		"The new description for the module",
+		"The new description for the plugin",
 	)
 	bufcli.BindStringPointer(
 		flagSet,
 		urlFlagName,
 		&f.URL,
-		"The new URL for the module",
-	)
-	flagSet.StringVar(
-		&f.DefaultLabel,
-		defaultLabelFlagName,
-		"",
-		"The label that commits are pushed to by default",
+		"The new URL for the plugin",
 	)
 }
 
@@ -90,11 +83,11 @@ func run(
 	container appext.Container,
 	flags *flags,
 ) error {
-	moduleFullName, err := bufparse.ParseFullName(container.Arg(0))
+	pluginFullName, err := bufparse.ParseFullName(container.Arg(0))
 	if err != nil {
 		return appcmd.WrapInvalidArgumentError(err)
 	}
-	visibility, err := bufcli.VisibilityFlagToVisibilityAllowUnspecified(flags.Visibility)
+	visibility, err := bufcli.VisibilityFlagToPluginVisibilityAllowUnspecified(flags.Visibility)
 	if err != nil {
 		return appcmd.WrapInvalidArgumentError(err)
 	}
@@ -102,44 +95,43 @@ func run(
 	if err != nil {
 		return err
 	}
-	moduleServiceClient := bufregistryapimodule.NewClientProvider(clientConfig).V1ModuleServiceClient(moduleFullName.Registry())
-	visibilityUpdate := &visibility
-	if visibility == modulev1.ModuleVisibility_MODULE_VISIBILITY_UNSPECIFIED {
-		visibilityUpdate = nil
+	var visibilityUpdate *pluginv1beta1.PluginVisibility
+	if visibility != pluginv1beta1.PluginVisibility_PLUGIN_VISIBILITY_UNSPECIFIED {
+		visibilityUpdate = &visibility
 	}
-	defaultLabelUpdate := &flags.DefaultLabel
-	if flags.DefaultLabel == "" {
-		defaultLabelUpdate = nil
-	}
-	if _, err := moduleServiceClient.UpdateModules(
-		ctx,
-		&connect.Request[modulev1.UpdateModulesRequest]{
-			Msg: &modulev1.UpdateModulesRequest{
-				Values: []*modulev1.UpdateModulesRequest_Value{
-					{
-						ModuleRef: &modulev1.ModuleRef{
-							Value: &modulev1.ModuleRef_Name_{
-								Name: &modulev1.ModuleRef_Name{
-									Owner:  moduleFullName.Owner(),
-									Module: moduleFullName.Name(),
-								},
+
+	pluginServiceClient := bufregistryapiplugin.NewClientProvider(clientConfig).
+		V1Beta1PluginServiceClient(pluginFullName.Registry())
+
+	pluginResponse, err := pluginServiceClient.UpdatePlugins(ctx, connect.NewRequest(
+		&pluginv1beta1.UpdatePluginsRequest{
+			Values: []*pluginv1beta1.UpdatePluginsRequest_Value{
+				{
+					PluginRef: &pluginv1beta1.PluginRef{
+						Value: &pluginv1beta1.PluginRef_Name_{
+							Name: &pluginv1beta1.PluginRef_Name{
+								Owner:  pluginFullName.Owner(),
+								Plugin: pluginFullName.Name(),
 							},
 						},
-						Description:      flags.Description,
-						Url:              flags.URL,
-						Visibility:       visibilityUpdate,
-						DefaultLabelName: defaultLabelUpdate,
 					},
+					Visibility: visibilityUpdate,
 				},
 			},
 		},
-	); err != nil {
+	))
+	if err != nil {
 		if connect.CodeOf(err) == connect.CodeNotFound {
 			return bufcli.NewModuleNotFoundError(container.Arg(0))
 		}
 		return err
 	}
-	if _, err := fmt.Fprintf(container.Stdout(), "Updated %s.\n", moduleFullName); err != nil {
+	plugins := pluginResponse.Msg.Plugins
+	if len(plugins) != 1 {
+		return syserror.Newf("unexpected number of plugins returned from server: %d", len(plugins))
+	}
+	_, err = fmt.Fprintf(container.Stdout(), "Updated %s.\n", pluginFullName)
+	if err != nil {
 		return syserror.Wrap(err)
 	}
 	return nil